Avoid triggering valgrind false positive
Valgrind memcheck produced a false positive for a memory leak in the
'subscribe_flaky_server' test case. It seems to be the use of fork()
and inlining which confuses valgrind.

This happens only at -O3 and -Os GCC optimization levels.

The patch works around the issue by preventing inlining of the
fork()-ing function.

Change-Id: I37b220a33ece5f688a0d68bc91b2a107d56b5fda