Introduce JSON file format and per-server TLS configuration
This patch introduces a new JSON-based domain file format, documented
in <paf.h>. The old domain file format, with the server addresses as a
set of newline-separated XCM addresses, is still supported.

The JSON format allows configuring certificate, private key and
trusted CAs on a per-TLS server basis. For this to work, libpaf needs
to be built and run with XCM API/ABI version 0.16 or later.

The test suite now runs the library with TLS, as well as TCP and UNIX
domain sockets. A Python script 'gencert.py' is used to generate the
required PKI structure.

Building against XCM API/ABI 0.13 is still supported, but such builds
will not allow configuring certificates and related files in the
domain files.

Resolves #14.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>