Use getentropy for random numbers
libpaf requires a high-quality random number generator (RNG). This
patch switch from reading /dev/urandom, to the getentropy() libc call.

The entropy source in the kernel is the same, but the special-purpose
function is a faster (although it rarely matters), and the code is
more compact.

The getentropy() function requires glibc 2.25, and the underlying
getrandom() syscall is available starting with Linux kernel 3.17.

Resolves #4.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>