Implement detachment timeout to handle unresponsive servers
The paf_detach() call now holds its promise to put an upper bound to
server detachment process. The bound is set to 500 ms + 1 ms for each
service published on the server. Applications wanting to close down
the context sooner that are free to do so.

Before this change, an unresponsive server could cause the detachment
to never success (and thus paf_process() to never return
PAF_ERR_DETACHED).

The previously #if 0-ed test case to verify this behavior is not put
into production.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>