Allow user per-service TTL configuration
This change introduces a new API call paf_set_ttl(), which allows
applications to modify default TTL of a particular service, after it
has been published.

The Python API (in libpaf.py) is updated accordingly.

The ABI version is now 0.1, and the library implementation version is
changed to 1.1.0.

This is purely a client library change. The Pathfinder protocol
supports (and has always supported) client-specified per-service TTLs.

Resolves #15.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>