Properly NUL terminate large log messages
Very large log messages weren't properly NUL terminated.

In some extreme cases this could results in garbled log entries and
potentially even crashes in case logging was enabled and, for example,
a service with very large properties were published via the C API.

This bug does not affect operation when logging is disabled (i.e. the
relevant LTTgn trace point is not enabled, and PAF_DEBUG is not set).

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>