Fix misleading connection establishment trace message
The library logged a just-initiated XCM connection as "established",
which is misleading. In the case XCM TLS or TCP is being used for
libpaf <-> server communication, xcm_connect() returning a XCM socket
does not mean the TCP and/or TLS-level handshake has finished, and
thus the XCM-level connection can't be said to be established.

This patch also remove the local address from the trace entry, since
it is usually NULL.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>