Avoid compiler warnings causing production build failures
Remove the -Werror default, since this flag does not make sense for
libpaf consumer-side production builds. The use of this compiler flag
might cause build failures, in particular if an older version of
libpaf is compiled with a newer (or otherwise different) compiler than
was used during libpaf development.

-Werror should still be used during development, and may still be
enabled by passing CFLAGS=-Werror to the configure script (even in
production builds).

The check-release.py script is updated to enable -Werror for its
builds.

Resolves #16.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>