Fix domain file read race
libpaf used the domain file's modification time as a means to detect
changes.

In a scenario with non-atomic domain file modifications, an unlucky
libpaf rescan may pick up an empty file, noting the mtime, and
clearing the server list for that domain. The domain file writer would
then go on to write the actual server list. Worse, the situation may
not resolve itself during the next rescan, since the mtime may not be
different than it was before.

Non-atomic here means that a writer uses in-place modification with
open()+write()+close() instead of rename() to replace the old domain
file with a new.

This patch discontinues the use of fstat() to try to detect changes.
The domain file is re-read and re-parsed for every rescan period.

Rescans are few and far between, so the original, failed, attempt at
optimizing rescans was not warranted.

Resolves #19.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>