Fix always active context fd bug
The timer fd used for connection retries was left active on successful
connection establishment. This in turn left the context's epoll fd to
be active, suggesting the application to call paf_process().

paf_process() wouldn't clear the timer, causing well-behaved
applications to consume excessive amount of CPU.

This bug was introduced in 1faf230.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>