Allow specifying the source address to use for a particular server
Add a new key "localAddress" to the server object in the JSON-based
domain file format. The address must be in XCM format.

If localAddress is set, the outgoing transport connections will be
originating from that address, for that server.

In regular XCM manner, the local address' port number number may be
set to 0, in which case the kernel will allocate a port number (from
the local port range).

Resolves #23.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>