Add support for certificate revocation verification
Allow configuring X.509 certificate revocation lists (CRLs) in the
domain file.

The tlsCrlFile parameter may only be present when libpaf is linked to
libxcm v1.9.0 or newer. When a CRL file is configured, CRL checking
will be enabled.

libpaf still supports older versions (>= v1.5.0) of XCM.

The test suite is updated with CRL-related integration tests.

To easily generate test CRLs, this patch bring in the newest version
of the 'gencert.py' tool from the XCM repo.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>