Handle terminal errors during server detachment
libpaf would restart the server link in case a terminal error (such as
a protocol violation) would occur, regardless of the state of the
link.

This may lead to strange behaviour where an application calls
paf_detach(), some sort of error occurs (e.g., a server violates the
Pathfinder protocol), and a new connection is established toward the
server, permanently abandoning the detachment process.

With this change, the detachment is considered completed in those
unusual cases where a terminal error occurs mid-process.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>