Add low-level Pathfinder protocol client API
Add an libpaf-internal module 'conn'. A conn object represent a
Pathfinder protocol connection, and has an API closely matching the
Pathfinder protocol-level primitives. It's lower level than the public
<paf.h>, but yet provides some abstraction over 'proto_ta'.

The new 'conn' module provides a complete Pathfinder protocol
implementation, and thus includes functionality such as Pathfinder
protocol-level "ping", listing subscriptions, clients and services on
the server - functionality that is not used by <paf.h> and the libpaf
library.

This patch also extends 'proto_ta' to support all protocol commands.

'conn' is currently not used by the libpaf library itself, but the
plan is to introduce it between 'link' and 'proto_ta'.

The primary aim for 'conn' is to facilitate a new command-line
debugging tool.

This patch also include a new test suite, testing 'conn'
against a Pathfinder server.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>