Add command-line Pathfinder client
Add a new interactive command-line program 'lpafc', which is a
reimplementation of the 'pafc' program of the Python-based Pathfinder
server software suite.

lpafc supports all pafc commands, using the same names and syntax. The
program output is also essentially the same, but may differ in some
details.

Command-line editing and completion is supported.

The server address must be specified as a command-line argument, as
lpafc does not support parsing Pathfinder domain files.

lpafc, just like pafc, is intended for debugging and exploration by a
human operator. It's not intended to be a machine-machine interface.

lpafc depends on GNU readline.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>