Improve certificate revocation test
Extend certificate revocation test with a scenario where the
certificate is revoked already at the point of the paf_attach() call.

Prior to this change, the tested scenario was the more elaborate one,
where the server certificate was revoked (and the client-side CRL
created) when the connection was already established.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>