Avoid crashing on protocol transaction failures
In case a command transaction failed due to some Pathfinder
protocol-level issue, the link was reset, but the loop that read the
original response causing the failure would go on to attempt to read
another message, on a now-closed-and-disposed-of XCM connection.

Beyond the fix, this patch also include a test case, reproducing the
issue.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>