Fix variadic function bug in protocol message generation
The variable argument list function proto_ta_produce_request() would
delegate work to other functions, erroneously passing a va_list
by-value, when the code following such calls assumed by-reference
semantics.

One of the symptoms of this bug is libpaf sending bogus version
information in the hello message. The bug has only been reproduced on
64-bit ARM. x86_64 builds seem to work. This difference is likely due
to differences in varargs-related ABI details between ARM and x86.

This bug was introduced in 1.0.4.

The patch makes the helpers macros instead of functions, to avoid the
hassle of dealing with subtle va_list semantics.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>