Avoid duplicate log prefix
A conn instance used a log prefix with client id, which doens't make
sense since such information is not uniquely tied to a connection, but
rather to a client.

After this patch, a connection does not add any to the supplied prefix
(i.e., it will reuse the lpafc or link-supplied prefix).

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>