Eliminate paf.client dependency
In the libpaf test suite, replace the use of the paf.client Python
module found in the 'paf' Python-based server repository with the
libpaf-local 'conn' module.

Now, the libpaf test suite only need *a* Pathfinder server, and can be
tested with only tpafd (for example), without any need to pull in the
'paf' repo as well.

The libpaf library itself was already independent of any particular
server implementation.

Resolves #28.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>