Add support Pathfinder protocol version 3
libpaf and lpafc is updated to support Pathfinder protocol version 3.

On version 3 connections, libpaf uses the new 'track' protocol command
to ensure that the server is alive.

The API documentation is updated with information libpaf liveness
checking, including information on the new PAF_IDLE_MIN and
PAF_IDLE_MAX environment variables.

lpafc now supports the additional information provided in the v3
'clients' command. A new lpafc 'track' command is added, matching the
Pathfinder protocol-level construct.

The tools and the library itself still support Pathfinder protocol
version 2 as well.

The application API <paf.h> is not affected by this change. However,
not adhering to the API contract and failing to calling paf_process()
when the context fd is active may cause connection(s) to being brought
down.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>