Fix bug server EOF causing link stuck in operational state
In case the server would shut down the connection using a normal
three-way handshake (casuing recv() to return 0 on the peer) after the
Pathfinder protocol level handshake had finished, the libpaf link
could be stuck in the operational state forever.

The bug only occured if errno happened to have the value EAGAIN going
in to the libpaf-internal conn_process() function.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>