Improve lpafc track command
The lpafc track command syntax is remodeled to fit a knob to turn
logging server queries on and off. Logging is now off by default.

If a low-TTL service is published, the server will send frequent track
queries, which makes the pafc CLI hard to use, in case each reply is
logged to the console.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>