Allow reducing client-advertised protocol version support
Allow the user to specify (within the limits of what the libpaf
software supports) minimum and/or maximum protocol versions allowed
during the protocol version negiotiation against a particular server.

Two new keys "minProtocolVersion" and "maxProtocolVersion" are added
to the server objects in the domain file.

The API documentation in <paf.h> is updated.

Test cases are added in the 'domain_conf', 'conn' and 'paf' test
suites. In particular, select 'paf' test cases are now run both with
any protocol version, and with the use of v2 (which all servers
support) being forced.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>