Avoid triggering idle timer in overload scenarios
libpaf only treated the server as known-to-be-alive when it received a
track query or a track reply notification.

There could be a situation where the server's track reply message was
head-of-line-blocked by a very large amount of notifications (or other
messages), causing libpaf to terminate the connection due to an idle
timeout, when it was the exact opposite of idle.

This only occurs either when the TTL is set very low, or when the
server sends messages at a very high rate indeed for a long period of
time.

With this patch, most server response message are treated as a sign of
life, and will post-poned both track query and reply timeouts.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>