Limit outstanding transactions
libpaf did not impose an upper limit for the number of concurrent,
unacknowledged protocol transactions toward the server.

"Unacknowledged" here means has-not-yet-received-a-complete-message
for publish transactions, and has-not-yet-received-an-accept-message
for subscribe transactions.

On protocol v3 connections, this causes issues since message queuing
in either the client, the server and (more unlikely) the network
causes delays for later transactions. In case such transactions are
track queries, the delay may cause dropped connections.

With this change, no more than 32 concurrent unacknowledged
transactions are allowed.

Resolves #31.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>