Tune idle timeouts
Change the default minimum max idle time from 2 to 4 seconds.

Change the idle query threshold from 75% of the max idle time, to 50%.

These changes are to reduce track query latency requirements, which in
turn improve robustness in situations when client, the network or the
server are not very responsive.

A small random component is added to the track query interval, to
avoid all clients connecting at the same time sending a query at the
same time, repeatedly. This in turn to avoid server-side load spikes.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>