Add initial handshake timeout
Introduce an upper bound for the amount of time the initial transport
layer (e.g., TCP and TLS) connection handshake and the Pathfinder
protocol hello transaction is allowed to take to complete. Timer
expiration trigger a connection restart.

The handshake timeout is set to the minimum idle time (e.g.,
PAF_IDLE_MIN, if set, or 4 seconds otherwise).

This change improves robustness in the face of certain "middle box"
failures. For example, it addresses a scenario where a TCP relay
sitting between the client and the server accepts a connection, but
never actually relay any traffic between the client and the server,
and where the reestablishment of a new connection works around this
"network" issue.

The test suite is extended to implement the above-mentioned scenario.

Resolves #29.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>