Improve subscription match state handling
This commit is a rewrite of the match module to more carefully track
per-server ("source") state, to be able to verify server behavior and
better handle certain race conditions in high availability (i.e.,
multiple-server) configurations.

With this change, libpaf will keep track of past subscription matches,
including those which are obsolete (e.g., which pertains to older
generations of the same service, or state related to an unpublished
service). Among other things, this allows libpaf to verify that any
server-induced subscription match state transitions are legal (i.e.,
apppeared [-> modified] -> disappeared).

A set of sequences of appeared-modified-disappeared notification
messages for a particular service in a particular subscription may
arrive out-of-order, provided multiple servers serve the same
domain. (This is not a bug, but a part of the protocol design.)

In case a service is very short-lived, the "appeared" message may
arrive at a particular client after a "disappeared" message have been
processed from some other server, for the same generation. In such a
scenario, libpaf prior to this commit would report the service being
matched (e.g., appeared) multiple times, even though it fact it was
just the same information taking different routes.

With this commit, libpaf maintains old ("stale") matches, so it's able
to filter out redundant information, and will in the above-mentioned
scenario only notify the application once.

To avoid an ever-growing memory footprint for long-lived
subscriptions, the matches are only kept around for ~30 s. In case an
"appeared" message from the short-lived-service scenario from above is
delayed longer, it will be treated as a new match.

libpaf now relies on its own timers only for unconnected orphaned
subscription matches (and not for all orphans). This approach avoids a
race between the remote, service-side timeout and the local,
client-side timeout, which in turn simplifies the libpaf logic
somewhat, especially since this version attempts more closely track
server behavior (and deviations from the protocol).

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>