Fix memory leak triggered by republishing stale services
The subscription match logic would leak a props object in the
following scenario:

1. A service matching a certain subscription becomes an orphan, and later
   stale (upon reaching the orphan timeout).
2. The service is republished.

This leak could also be triggered by a client publishing, explicitly
unpublishing and then again publishing the same service.

The leak was detected by the 'sub:provider_client_afk' test case when
run in valgrind, but for some reason this error went unnoticed for
a long time.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>