Fix bug leaving stale service or subscription relays
In case a service or subscription was removed from a context by the
application (e.g., with paf_unpublish()) while the object was not yet
synced to the server, the service or subscription relay was left on
the link.

This could cause libpaf to later attempt to sync objects no longer in
libpaf's internal database, which caused it to dereference a NULL
pointer.

This patch fixes the bug.

In addition, a new test case is introduced, which publishes,
unpublishes, subscribes and unsubscribes in a random pattern, with the
servers going up and down in a random pattern as well. This test was
able to reproduce this bug.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>