Assure libreadline is not linked to libpaf on RHEL
On some systems (e.g., RHEL), -Wl,--as-needed is not a default linker
flag in their toolchains. libpaf automake is set up in such a manner,
that all programs and shared objects built are linked toward all
libraries, and the mentioned linker flag is dependen upon to make the
linker prune unused library dependencies.

This patch makes sure this flag is always set.

With this change, non-GNU compatible linkers can no longer be used in
libpaf builds. That is likely more of a feature than a bug, since
otherwise the end user may end up with spurious libraries dependencies
in the libpaf shared object.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>