Explicitly list library dependencies
Explicitly specify shared library dependencies for each program or
shared object built by the libpaf build system.

This is to make dependencies clear, and also to address issues with
readline needlessly being linked to libpaf.so on systems which does
not use pass --as-needed to the C compiler by default.

This patch backs out the previous patch, which attempted (and failed)
to always pass --as-needed to the linker.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>